Interview_Question_Avyott/requirements.txt
streamlit 
pandas 
pytesseract 
pillow 
pdfplumber 
fpdf